Move shared code out of JacocoPlugin
