+++ build.sbt
@@ -3,9 +3,8 @@ lazy val root = (project in file(".")).enablePlugins(BuildInfoPlugin).settings(
   name := "jacoco4sbt",
   organization := "de.johoop",
   version := "3.0.0-SNAPSHOT",
-  scalaVersion := "2.10.6",
 
-  crossSbtVersions := Vector("0.13.16", "1.0.0"),
+  crossSbtVersions := Vector("0.13.16", "1.0.1"),
 
   sbtPlugin := true,
